Rearrange; add comments
