+++ src/components/Conversation/MessageFeed.tsx
@@ -1,4 +1,4 @@
-import { Fragment, HTMLAttributes, useEffect, useLayoutEffect, useMemo, useRef } from 'react'
+import { Fragment, HTMLAttributes, useEffect, useLayoutEffect, useRef } from 'react'
 import tw from 'twin.macro'
 import { IMessage, IResend } from '$/features/message/types'
 import { useWalletAccount } from '$/features/wallet/hooks'
@@ -6,8 +6,8 @@ import isSameAddress from '$/utils/isSameAddress'
 import Message from './Message'
 import { useSelectedRoomId } from '$/features/room/hooks'
 import MessageGroupLabel from '$/components/Conversation/MessageGroupLabel'
-import getBeginningOfDay from '$/utils/getBeginningOfDay'
 import Text from '$/components/Text'
+import useMessageGroups from '$/hooks/useMessageGroups'
 
 type Props = {
     messages?: IMessage[]
@@ -15,82 +15,6 @@ type Props = {
 }
 
 const AutoScrollTolerance = 5
-
-interface Manifest {
-    messages: IMessage[]
-    newDay: boolean
-}
-
-interface Manifests {
-    [timestamp: string]: Manifest
-}
-
-interface Group extends Manifest {
-    timestamp: number
-}
-
-function useMessageGroups(messages: IMessage[], resends: IResend[]): Group[] {
-    return useMemo(() => {
-        const manifests: Manifests = {}
-
-        resends.forEach((r) => {
-            manifests[r.beginningOfDay] = {
-                messages: [],
-                newDay: true,
-            }
-        })
-
-        let previousCreatedAt: undefined | number = undefined
-
-        let rangeTimestamp: undefined | number = undefined
-
-        messages.forEach((m) => {
-            if (typeof m.createdAt !== 'number') {
-                return []
-            }
-
-            let startsDay = true
-
-            let startsRange = true
-
-            const bod = getBeginningOfDay(m.createdAt)
-
-            if (typeof previousCreatedAt === 'number') {
-                startsDay = bod !== getBeginningOfDay(previousCreatedAt)
-
-                // Group messages that happen within 15 minutes from each other.
-                startsRange = !startsDay && previousCreatedAt + 900000 <= m.createdAt
-            }
-
-            previousCreatedAt = m.createdAt
-
-            if (manifests[bod]?.messages?.length === 0) {
-                delete manifests[bod]
-            }
-
-            if (startsRange || startsDay) {
-                rangeTimestamp = m.createdAt
-            }
-
-            if (typeof rangeTimestamp !== 'number') {
-                throw new Error('This should not happen. Please report if it did.')
-            }
-
-            if (!manifests[rangeTimestamp]) {
-                manifests[rangeTimestamp] = {
-                    messages: [],
-                    newDay: startsDay,
-                }
-            }
-
-            manifests[rangeTimestamp].messages.push(m)
-        })
-
-        return Object.keys(manifests)
-            .sort()
-            .map((ts) => ({ timestamp: Number(ts), ...manifests[ts] }))
-    }, [messages, resends])
-}
 
 export default function MessageFeed({ messages = [], resends = [] }: Props) {
     const rootRef = useRef<HTMLDivElement>(null)
